requirements.txt
streamlit
pandas
scikit-learn
numpy
openpyxl
plotly
reportlab
google-genai
